+++ SCAS/src/java/dao/FormularioDAO.java
@@ -56,46 +56,18 @@ public class FormularioDAO {
         return formulario;
     }
     
-    public void gravar(Formulario formulario){
+    public void operacao(Formulario formulario, String operacao){
         EntityManager em = PersistenceUtil.getEntityManager();
         EntityTransaction tx = em.getTransaction();
         try {
             tx.begin();
-            em.persist(formulario);
-            tx.commit();
-        } catch (Exception e) {
-            if (tx != null && tx.isActive()) {
-                tx.rollback();
-            }
-            throw new RuntimeException(e);
-        } finally {
-            PersistenceUtil.close(em);
-        }
-    }
-    
-    public void alterar(Formulario formulario){
-        EntityManager em = PersistenceUtil.getEntityManager();
-        EntityTransaction tx = em.getTransaction();
-        try {
-            tx.begin();
-            em.merge(formulario);
-            tx.commit();
-        } catch (Exception e) {
-            if (tx != null && tx.isActive()) {
-                tx.rollback();
-            }
-            throw new RuntimeException(e);
-        } finally {
-            PersistenceUtil.close(em);
-        }
-    }
- 
-    public void excluir(Formulario formulario){
-        EntityManager em = PersistenceUtil.getEntityManager();
-        EntityTransaction tx = em.getTransaction();
-        try {
-            tx.begin();
-            em.remove(em.getReference(Formulario.class, formulario.getCodFormulario()));
+            if(operacao.equals("gravar")){
+                em.persist(formulario);
+            } else if(operacao.equals("alterar")){
+                em.merge(formulario);
+            } else if(operacao.equals("excluir")){
+                em.remove(em.getReference(Formulario.class, formulario.getCodFormulario()));
+            }            
             tx.commit();
         } catch (Exception e) {
             if (tx != null && tx.isActive()) {
