+++ SCAS/src/java/dao/RendaDAO.java
@@ -56,46 +56,18 @@ public class RendaDAO {
         return renda;
     }
 
-    public void gravar(Renda renda){
+    public void operacao(Renda renda, String operacao){
         EntityManager em = PersistenceUtil.getEntityManager();
         EntityTransaction tx = em.getTransaction();
         try {
             tx.begin();
-            em.persist(renda);
-            tx.commit();
-        } catch (Exception e) {
-            if (tx != null && tx.isActive()) {
-                tx.rollback();
-            }
-            throw new RuntimeException(e);
-        } finally {
-            PersistenceUtil.close(em);
-        }
-    }
-
-    public void alterar(Renda renda){
-        EntityManager em = PersistenceUtil.getEntityManager();
-        EntityTransaction tx = em.getTransaction();
-        try {
-            tx.begin();
-            em.merge(renda);
-            tx.commit();
-        } catch (Exception e) {
-            if (tx != null && tx.isActive()) {
-                tx.rollback();
-            }
-            throw new RuntimeException(e);
-        } finally {
-            PersistenceUtil.close(em);
-        }
-    }
-
-    public void excluir(Renda renda){
-        EntityManager em = PersistenceUtil.getEntityManager();
-        EntityTransaction tx = em.getTransaction();
-        try {
-            tx.begin();
-            em.remove(em.getReference(Renda.class, renda.getCodRenda()));
+            if(operacao.equals("gravar")){
+                em.persist(renda);
+            } else if(operacao.equals("alterar")){
+                em.merge(renda);
+            } else if(operacao.equals("excluir")){
+                em.remove(em.getReference(Renda.class, renda.getCodRenda()));
+            }            
             tx.commit();
         } catch (Exception e) {
             if (tx != null && tx.isActive()) {
